feat: add file creation and processing functions with database integration
